requirements.txt
google-generativeai
streamlit
pytz